+++ meta/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the Orbital Trading project will be documented in this file by the user, only. Gemini is never to modify or provide updates to this file, and must always treat this file as read-only. The user will always manually update this file.
 
+## [33.56] - 2026-1-1
+-  updated Water Ice assets (5)
+-  added commodity card background functionality
+-  added commodity card asset cycling
+-  updated general asset cycling each day
+-  added screen redraw to World & Time debug
+
 ## [33.50] - 2025-12-31
 -  updated Thalassodromeus assets (13)
 -  updated Ouroboros assets (13)